requirements.txt
tensorflow
tensorflow-hub
numpy
pillow